tests/requirements.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
pyyaml>=6.0.0
playwright>=1.44.0
